Jersey version update. Fix client initialization.

+++ client/src/main/java/org/gluu/oxtrust/api/test/saml/SamlTestScenary.java
@@ -43,31 +43,38 @@ public class SamlTestScenary {
     public void run() throws APITestException, OxTrustAPIException {
         TrustRelationshipClient samlClient = client.getTrustRelationshipClient();
         
-        GluuSAMLTrustRelationship trGenerated = generateRandomeSingleTrustRelationship();
-        
-        // test create()
-        String inum = samlClient.create(trGenerated);
-        
-        // test read()
-        GluuSAMLTrustRelationship trReaded = samlClient.read(inum);
-        //TODO: compare etities
-        trReaded.setDescription("description changed");
-        
-        // test update()
-        samlClient.update(trReaded, inum);
-        
-        // test list()
         List<SAMLTrustRelationshipShort> trustRelationships = samlClient.list();
-        if (!checkListForTrustRelationship(trustRelationships, inum))
-            throw new APITestException("TrustRelationship really not saved");
         
-        // test delete()
-        samlClient.delete(inum);
-        trustRelationships = samlClient.list();
-        if (checkListForTrustRelationship(trustRelationships, inum))
-            throw new APITestException("TrustRelationship really not deleted");
+        // prevent server data corrupton - work with empty dataset
+        if (trustRelationships.isEmpty()) {
+            GluuSAMLTrustRelationship trGenerated = generateRandomeSingleTrustRelationship();
+
+            // test create()
+            String inum = samlClient.create(trGenerated);
+
+            // test read()
+            GluuSAMLTrustRelationship trReaded = samlClient.read(inum);
+            //TODO: compare etities
+            trReaded.setDescription("description changed");
+
+            // test update()
+            samlClient.update(trReaded, inum);
+            
+            // test list()
+            trustRelationships = samlClient.list();
+            if (!checkListForTrustRelationship(trustRelationships, inum))
+                throw new APITestException("TrustRelationship really not saved");
+    //        
+    //        // test delete()
+            samlClient.delete(inum);
+            trustRelationships = samlClient.list();
+            if (checkListForTrustRelationship(trustRelationships, inum))
+                throw new APITestException("TrustRelationship really not deleted");
+
+            //TODO: all API calls
+        }
         
-        //TODO: all API calls
+        
     }
     
     private GluuSAMLTrustRelationship generateRandomeSingleTrustRelationship() {
